--- DataMigrationUsingFetchXml/Model/FormData.cs
@@ -1,18 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace DataMigrationUsingFetchXml.Model
-{
-    public sealed class FormData
-    {
-        public Label LblInfo { get; set; }
-        public Label LblTitle { get; set; }
-        public Label LblError { get; set; }
-
-        public FormData(Label lblInfo, Label lblTitle, Label lblError)
-        {
-            LblInfo = lblInfo;
-            LblTitle = lblTitle;
-            LblError = lblError;
-        }
-    }
-}
